fix: several bugs fixed after testing

+++ backend/app.py
@@ -44,16 +44,21 @@ def simulation(simulationId):
         return make_response(jsonify(simulator.loadSimulationById(simulationId)), 200) #give back the selected sim
     
     elif request.method == 'DELETE':
-        simulator.deleteSimulationById(simulationId) #delete a selected sim
-        return make_response(f"Simulation '{simulationId}' was deleted successfully", 200)
-
+        if DatabaseHandler.selectMachineState(simulationId):
+            simulator.deleteSimulationById(simulationId) #delete a selected sim
+            return make_response(f"Simulation '{simulationId}' was deleted successfully", 200)
+        else:
+            return make_response(f"Simulation '{simulationId}' does not exist", 400)
 # ------------------------------------------------------------------------------
 # Protocol
 @app.route('/api/simulations/protocol', methods=['PATCH'])
 def simulationsProtocol():
   if request.method == 'PATCH':
-    simulator.setProtocol(request.form['protocol'])
-    return make_response(f"Protocol '{request.form['protocol']}' was set successfully", 200)
+    if request.form['protocol'] == 'OPCUA' or request.form['protocol'] == 'Modbus/TCP' or request.form['protocol'] == 'None':
+        simulator.setProtocol(request.form['protocol'])
+        return make_response(f"Protocol '{request.form['protocol']}' was set successfully", 200)
+    else:
+        return make_response(f"Protocol '{request.form['protocol']}' is not supported", 400)
 
 # ------------------------------------------------------------------------------
 # Machine
@@ -61,7 +66,7 @@ def simulationsProtocol():
 def machines():
     if request.method == 'GET':
         simulator.updateSimulation(datetime.now())
-        return make_response(jsonify(simulator.simulatedMachine.getMachineStateSnapshot())) #give back the current machine state
+        return make_response(jsonify(simulator.simulatedMachine.getMachineStateSnapshot()), 200) #give back the current machine state
     elif request.method == 'PATCH':
         simulator.updateMachineStateParameters(request.get_json())
         return make_response("Machine state parameters updated successfully", 200)  #change parameter(s) in the machine state
@@ -104,13 +109,14 @@ def error():
 
 # ------------------------------------------------------------------------------
 # Programs
-@app.route('/api/simulations/machine/programs')
+@app.route('/api/simulations/machine/programs', methods=['GET'])
 def allPrograms():
-    listOfPrograms = simulator.getPrograms()
-    data = {"programs": []}
-    for program in listOfPrograms:
-        data["programs"].append(program.toJSON())
-    return make_response(jsonify(data), 200) #list of all programs
+    if request.method == 'GET':
+        listOfPrograms = simulator.getPrograms()
+        data = {"programs": []}
+        for program in listOfPrograms:
+            data["programs"].append(program.toJSON())
+        return make_response(jsonify(data), 200) #list of all programs
 
 # ------------------------------------------------------------------------------
 # Current Program
